# packages/api-client/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Medagil API
  version: 0.2.0
  description: >
    Medagil 医学科研写作平台 API 规范（MVP 阶段）。
    - 用户端与管理端 REST API，路径与 tag 区分
    - 内部接口（internal tag）用于 api-service 与 ai-service / 其它服务间调用
servers:
  - url: ''
    description: 运行时由各 app 注入 baseUrl（如 process.env.NEXT_PUBLIC_API_URL）
...
paths:
...
  /api/v1/me:
    get:
      tags: [user]
      summary: 当前用户信息
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: 未登录

    patch:
      tags: [user]
      summary: 更新当前用户资料与偏好
      operationId: updateMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: 未登录
  /api/v1/projects:
    get:
      tags: [user]
      summary: 项目列表
      operationId: getProjects
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'

    post:
      tags: [user]
      summary: 创建项目
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: 参数错误
...
  /api/v1/tasks:
    get:
      tags: [user]
      summary: 任务列表
      operationId: getTasks
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: projectId
          in: query
          schema:
            type: string
        - name: agentKey
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, succeeded, failed]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'

    post:
      tags: [user]
      summary: 新建任务
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
  /api/v1/tasks/{taskId}:
    get:
      tags: [user]
      summary: 任务详情
      operationId: getTaskById
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
        '404':
          description: 未找到
...
  /api/v1/admin/dashboard/stats:
    get:
      tags: [admin]
      summary: 数据看板统计
      operationId: getDashboardStats
      security:
        - adminBearer: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          description: 未登录
        '403':
          description: 无权限
  /api/v1/admin/users:
    get:
      tags: [admin]
      summary: 用户列表（管理端）
      operationId: getAdminUsers
      security:
        - adminBearer: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
...
  /api/v1/admin/tasks:
    get:
      tags: [admin]
      summary: 全平台任务列表
      operationId: getAdminTasks
      security:
        - adminBearer: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: userId
          in: query
          schema:
            type: string
        - name: agentKey
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
  /api/v1/admin/tasks/{taskId}:
    get:
      tags: [admin]
      summary: 任务详情（管理端）
      operationId: getAdminTaskById
      security:
        - adminBearer: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
  /api/v1/admin/orders:
    get:
      tags: [admin]
      summary: 订单列表
      operationId: getAdminOrders
      security:
        - adminBearer: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderListResponse'
  /api/v1/admin/orders/{orderId}:
    get:
      tags: [admin]
      summary: 订单详情
      operationId: getAdminOrderById
      security:
        - adminBearer: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrder'
  /api/v1/admin/orders/{orderId}/refund:
    post:
      tags: [admin]
      summary: 退款
      operationId: refundOrder
      security:
        - adminBearer: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 已触发退款
  /api/v1/admin/knowledge/sources:
    get:
      tags: [admin]
      summary: 知识源列表
      operationId: getKnowledgeSources
      security:
        - adminBearer: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeSourceListResponse'

    post:
      tags: [admin]
      summary: 创建或上传知识源
      operationId: createKnowledgeSource
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeSource'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeSource'

  # ---------- 内部接口 (tag: internal) ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 用户端 Token
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 管理端管理员 Token
  schemas:
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        nickname:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        plan:
          type: string
          enum: [free, pro, max]
        credits:
          type: integer
          nullable: true
    UpdateUserProfileRequest:
      type: object
      properties:
        nickname:
          type: string
        avatar:
          type: string
        preferences:
          type: object
        profile:
          type: object
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
          format: date-time
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    ProjectListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectSummary'
        total:
          type: integer
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        globalInstruction:
          type: string
...
    Task:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
          nullable: true
        agentKey:
          type: string
        status:
          type: string
          enum: [pending, running, succeeded, failed]
        createdAt:
          type: string
          format: date-time
    TaskListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        total:
          type: integer
    TaskDetail:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
          nullable: true
        agentKey:
          type: string
        status:
          type: string
        inputText:
          type: string
          nullable: true
        fileIds:
          type: array
          items:
            type: string
        resultPreview:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateTaskRequest:
      type: object
      required: [agentKey]
      properties:
        projectId:
          type: string
        type:
          type: string
          enum: [text, file, advanced, speech]
        agentKey:
          type: string
        inputText:
          type: string
        fileIds:
          type: array
          items:
            type: string
        advancedConfig:
          type: object
...
    DashboardStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalOrders:
          type: integer
        mrr:
          type: number
    AdminUserListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserProfile'
        total:
          type: integer
    AdminOrder:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        planId:
          type: string
        status:
          type: string
        amountCents:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminOrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminOrder'
        total:
          type: integer
...
    KnowledgeSource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    KnowledgeSourceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeSource'
        total:
          type: integer
...