# packages/api-client/src/admin/model/taskDetail.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Medagil API
 * Medagil 医学科研写作平台 API 规范（MVP 阶段）。 - 用户端与管理端 REST API，路径与 tag 区分 - 内部接口（internal tag）用于 api-service 与 ai-service / 其它服务间调用

 * OpenAPI spec version: 0.2.0
 */

export interface TaskDetail {
  id?: string;
  /** @nullable */
  projectId?: string | null;
  agentKey?: string;
  status?: string;
  /** @nullable */
  inputText?: string | null;
  fileIds?: string[];
  /** @nullable */
  resultPreview?: string | null;
  createdAt?: string;
}