# packages/api-client/src/user/model/knowledgeSourceListResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Medagil API
 * Medagil 医学科研写作平台 API 规范（MVP 阶段）。 - 用户端与管理端 REST API，路径与 tag 区分 - 内部接口（internal tag）用于 api-service 与 ai-service / 其它服务间调用

 * OpenAPI spec version: 0.2.0
 */
import type { KnowledgeSource } from './knowledgeSource';

export interface KnowledgeSourceListResponse {
  items?: KnowledgeSource[];
  total?: number;
}